+++ object-construction-checker/src/main/java/org/checkerframework/checker/objectconstruction/ObjectConstructionChecker.java
@@ -1,4 +1,4 @@
-package org.checkerframework.checker.builder;
+package org.checkerframework.checker.objectconstruction;
 
 import java.util.LinkedHashSet;
 import org.checkerframework.checker.returnsrcvr.ReturnsRcvrChecker;
