+++ object-construction-checker/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.checkerframework.checker.builder.ObjectConstructionChecker
+org.checkerframework.checker.objectconstruction.ObjectConstructionChecker
