+++ object-construction-checker/stubs/DescribeImages.astub
@@ -1,6 +1,6 @@
 package com.amazonaws.services.ec2;
 
-import org.checkerframework.checker.builder.qual.*;
+import org.checkerframework.checker.objectconstruction.qual.*;
 import org.checkerframework.checker.returnsrcvr.qual.*;
 
 class AmazonEC2 {
