+++ object-construction-checker/tests/README.md
@@ -1,4 +1,4 @@
-This directory contains the various test suites for the unsafe builder checker.
+This directory contains the various test suites for the object construction checker.
 
 Each test suite is meant to reflect a particular application domain. The test suites are:
 * "basic": generic tests to make sure the typechecker is working as intended. These don't reflect an application domain.
