+++ object-construction-checker/tests/autovalue/GuavaImmutablePropBuilder.java
@@ -1,5 +1,5 @@
 import com.google.auto.value.AutoValue;
-import org.checkerframework.checker.builder.qual.*;
+import org.checkerframework.checker.objectconstruction.qual.*;
 import org.checkerframework.checker.nullness.qual.*;
 
 import com.google.common.collect.ImmutableList;
