+++ object-construction-checker/tests/basic/Subtyping.java
@@ -1,4 +1,4 @@
-import org.checkerframework.checker.builder.qual.*;
+import org.checkerframework.checker.objectconstruction.qual.*;
 
 // basic subtyping checks
 class Subtyping {
